+++ R/memory_check.R
@@ -11,6 +11,8 @@ fit.memory <- function(x, n=1){
   # x rasters will be generated in this function, let's see if there is enough memory in the user's pc
   sink(nullfile())    # suppress output
   mi <- terra::mem_info(x, n)[5] != 0 # proc in memory = T TRUE means that it fits in the pc's memory, so you wouldn't have to use temporary files
-  sink()
+  names(mi) <- NULL
+  sink(file = NULL)
+  # sink(file = NULL)
   return(mi)
 }